2nd Add Attachments button
A second Add Attachments button is rendered with the metabox but with the `hide` class by default. I've added appropriate handlers to bind it to launch the Add Media frame and show it by removing the `hide` class on page load (if there are any attachments already associated with the post) or in the callback after selecting an attachment in the frame.